Update TASK 4(PASSWORD GENERATOR)
Title: Password Generator in Python

Description:
This Python project implements a simple password generator application that allows users to specify the desired length of the password. The password generator utilizes the random module to generate a combination of random characters including lowercase letters, uppercase letters, digits, and punctuation symbols.

Key features of the password generator include:

User Input: Users are prompted to specify the desired length of the password.
Generate Password: The application generates a random password of the specified length using a combination of characters.
Display Password: The generated password is displayed on the screen for the user to use.
This project is suitable for beginners looking to practice Python programming and for anyone in need of a simple tool to generate strong and random passwords. Feel free to contribute, provide feedback, or customize the project according to your needs.